Video-downloader/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
yt-dlp
ffmpeg-python
yt_dlp
python-dotenv
requests